Move files to the utils/ dir
